// web/src/utils/backend-client/types/entitlement.ts
/**
 * Generated by orval v6.10.3 🍺
 * Do not edit manually.
 * Common Fate
 * Common Fate API
 * OpenAPI spec version: 1.0
 */
import type { TargetSchema } from './targetSchema';
import type { TargetGroupFrom } from './targetGroupFrom';

export interface Entitlement {
  Schema: TargetSchema;
  Kind: TargetGroupFrom;
}